Remove 30 characters restriction for project names
There are other places containing this restriction, too.
But as our Mantis database is not affected by those and
as I do not know why those restrictions have been introduced
in the first place and as I do not know about the side
effects if they are all removed and as time is short I leave
them in place.